+++ packages/list/src/group-title.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="weui-cells__title">新消息提示跟摘要信息后，统一在列表右侧</div>
+  <div class="weui-cells__title">{{ title }}</div>
 </template>
 <script>
 export default {